[MicroBlaze] minor class re-namings and method abstractions; test fixes

+++ MicroBlaze/src/org/specs/MicroBlaze/instruction/MicroBlazeInstruction.java
@@ -122,7 +122,7 @@ public class MicroBlazeInstruction extends AInstruction {
      */
     private MicroBlazeInstruction(String address, String instruction, InstructionData idata,
             MicroBlazeAsmFieldData fieldData, InstructionProperties props) {
-        super(Long.parseLong(address, 16), instruction, idata, props);
+        super(Integer.parseInt(address, 16), instruction, idata, props);
         this.fieldData = fieldData;
     }
